moved util to internal
