+++ services/gateway/handler/account/handler.go
@@ -7,9 +7,9 @@ import (
 	"github.com/google/uuid"
 	"github.com/idlab-discover/kebeng/common/cerror"
 	"github.com/idlab-discover/kebeng/services/gateway/handler/auth"
-	"github.com/idlab-discover/kebeng/services/gateway/handler/util"
-	storepb "github.com/idlab-discover/kebeng/services/store/proto"
+	"github.com/idlab-discover/kebeng/services/gateway/handler/internal/util"
 	"github.com/idlab-discover/kebeng/services/gateway/internal/model"
+	storepb "github.com/idlab-discover/kebeng/services/store/proto"
 	"gopkg.in/macaroon.v2"
 )
