+++ services/gateway/handler/auth/handler.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/idlab-discover/kebeng/common/cerror"
-	"github.com/idlab-discover/kebeng/services/gateway/handler/util"
-	storepb "github.com/idlab-discover/kebeng/services/store/proto"
+	"github.com/idlab-discover/kebeng/services/gateway/handler/internal/util"
 	"github.com/idlab-discover/kebeng/services/gateway/internal/model"
+	storepb "github.com/idlab-discover/kebeng/services/store/proto"
 )
 
 type Handler struct {
