+++ services/gateway/handler/main.go
@@ -6,8 +6,8 @@ import (
 	assertionClient "github.com/idlab-discover/kebeng/services/assertion/client"
 	account "github.com/idlab-discover/kebeng/services/gateway/handler/account"
 	auth "github.com/idlab-discover/kebeng/services/gateway/handler/auth"
+	"github.com/idlab-discover/kebeng/services/gateway/handler/internal/util"
 	snap "github.com/idlab-discover/kebeng/services/gateway/handler/snap"
-	"github.com/idlab-discover/kebeng/services/gateway/handler/util"
 	"github.com/idlab-discover/kebeng/services/gateway/internal/config"
 	"github.com/idlab-discover/kebeng/services/gateway/internal/middleware"
 	storeClient "github.com/idlab-discover/kebeng/services/store/client"
