+++ services/gateway/internal/util/util.go
